fixes to lookups and functions
Fixed typo in lookups.py in Table lifetime.  Also Z ranges for lifetime
are 0.001 and 0.02 in Schaller et al and not as they were quoted in
lookups.py.  Edited functions.py to read in low metal lifetime at
Z<0.008 and high metal lifetime at Z>0.008 since this is between 0.001
and 0.02